Fix issue with functions being exported in target lib/functions.dart
Update test/hello to fully exercise this pattern
